Version 0.9
Changes to filestructure a little bit.

Relocated declarations that omitted definitions for plugins to define INTO the plugins themselves.
Reason: It worked, but it was unnecessary. It complicated things when it didn't need to. Plugins will expand on the functionality it derives from, not complete it. The system shouldn't be a puzzle with pieces scattered about.

Fixed issue with model finalization crashing. Plugins link against core, but aren't core. They need glew to be initialized in them as well apparently. Plugins now have their own Initialization call just like Core does.