Version: 0.10
Combined all the prerequisites into a single build for now.
I need to focus on the core functionality before I start to design plugins and what dependencies can be swapped out.

I've opted to static link as many dependencies as possible, but I believe this should still work with the dynamic versions if anyone wants to use them and include the dlls in their own projects.