Version: 0.13
Added preliminary support for preference changing event callbacks.
Preference manager has a map of vectors, each vector corresponding to a specific preference. Those vectors hold a pointer to a callback function. Anything in the program can add a callback function pointer to the appropriate vector  by calling "addPreferenceCallback() with the preference key and the function.

Testing shows that this works, but I haven't implemented anything in the core library that uses it, though I do have a use for it (such as window resizing and post processing effects)