Version: 0.14
Beginning to add Entities. First one added: Prop
There were occasions in the past where some entities didn't need location information, or needed to specify a specific way of storing their location info (like physics props), so I made a separate Transform struct which entities can optionally use. Prop includes it as a member variable.

To Do: Standardize the transform buffer for the renderable types in a way that is intuitive to the end user.
The camera buffer should be consistent across all shaders (and to do: the engine buffer), but the transform buffer needs to be render type specific, as models need bone transforms but terrain and brushes don't.