Version: 0.20
A lot of changes, but no noticeable difference yet to the end user. Almost complete.

Implemented Shadow map manager, and also a visibility manager.

Cameras automatically add themselves to the visibility manager and it computes visibile objects in a separate thread.

Shadows not actually in yet, but I'm 95% of the way there now.