Version: 0.24
More entity-component systems stuff. Still not completely up and running.

Entities are stored in a map in the EntityFactory, the key is a char* of entity name, and the value is a vector of entity pointers.

Components are stored similarly in the ComponentFactory.

In the next commit we will clear out all the entity and component #define for type id's, get rid of the geometry manager, and get rid of the light manager. The component manager can retrieve a list of components grouped by their char* name anyway.