Version: 0.33
Starting to replace namespaces with systems

Engine is abstracted to its own class, though multiple instances aren't yet working in tandom, but this is planned

Modified configuration system so that a config asset requires an input list of strings to compare against when reading from a file, so that it can be used in different ways.
Now I have config.cfg and binds.cfg files.

User input currently works, but nothing actually uses that yet. I need to implement a client system.