Version: 0.39
Replaced namespace World_Manager with System_World. The world creates and owns entity and component factories. These are now classes, instead of namespaces.
Also changed the ECSmessanger from being a namespace to a class.

The messanger needs access to the factories, and so gets constructed in the World with pointers to the world's factories. In visa versa, the factories need references to the messangers (and to the other factory) to pass down to their objects. The component factory needs the reference to the messanger because components need the reference to the messanger (to pass messages to their entity parents). The entity factory needs the reference to the messange for a similar reason so that its entities can pass messages to their children. Entities also need a reference to the component factory so they can dictate how to create their own components.