Version: 0.43
Added missing code to shadowmap system.
Removed vestigial shadowmap system.

Made it possible for components to have access to the Engine_Package in their constructors. This way they can query it for any systems they may need access to, such as lights needing the shadowmap system. This same pattern can be used (if needed, not currently implemented) for entities, since their parent factory just passes the reference on to them.