Version: 0.46
Migrating Asset creation to make use of the secondary context, so that not only their initialization can be performed on a secondary thread (as was already the case) but their finalization involving interacting with the OpenGL can be also done in a separate thread.

To this end, I've started off with the more difficult asset : bindless textures (As they rely on an extension). Next comit I will add more.
The new asset manager is temporarly named "Asset_Managera", and is a singleton.
I could in theory have initialization+finalization in 1 function, however the slowest part of asset loading is the reading from disk.
To that end, the new asset manager has a variable number of worker threads.