Version: 0.48
Model assets now use the new asset manager, and can load into the scene in parallel.
VAO's however are context dependant and can't be created on a separate context.
Though, they can be created per component or per thing that needs to use them, so this is the work around I chose.

In the case of the animated model component, it keeps track of when it changes model and sets a flag reminding itself that it needs to update its vao with the updated buffers. Even though the buffers are independant of the vao and can change on their own, if the model is swapped out with another one, it will have different buffers and ergo we to set bind those buffers in our VAO.

The check is performed on draw and also in the update method. Both of these methods should only ever be called from the rendering thread anyway.