Version: 0.53
Updated (and re-implemented, b/c they did nothing) cubemap assets to make use of the new asset manager system.

I tried to do a crude skybox but I can't get it to extrude around the depthmap correctly yet, so there is some blending issues.
But that doesn't matter, because I will handle skyboxes in some fashion properly in the future.