Version 0.54
Updated Collider Asset to use the new asset manager system.

Note: physics simulation not implemented, so changes likely will have to be made to this class. Further, the asset loads on a secondary thread and would benefit from a system in a similar way to models/primitives by signaling to some parent when it has finished.