Version: 0.56
Fixed mutex lock crash when loading default asset
Added asset notification system which allows an asset to notify observers when it has finished loading.

Things that want to use this system define and create their own observers that extend the Asset_Observer, and store it. They overload the Notify function, and the logic gets entirely carried out in there.
Observers tell the asset to add itself to the assets list on construction, and remove it on destruction. Observer destruction is the responsibility of the one who created it only. For safety sake, they should store shared pointers instead of normal pointers, in case a notification event is created but the asset or observer is deleted while the notification hasn't been processed yet.