Version: 0.63
Materials\Models\ folder using a similar file structure for the model's location, changing the file extension to .mat.
All material texture file paths are assumed to be relative to the Textures\Environment\ folder. NOT enforcing the model folder within there in case a particular model wants to share an entire skin or a particular texture file with another.