Version: 0.64
Improved material loading logic.

Models can choose their own textures for their material. If those images don't exist, the material attempts to load a default one of the appropriate category. If that doesn't exist, it uses a hardcoded value for the appropriate category.
If the model doesn't specify its own textures, the system looks for an appropriate .mat file, and attempts to load the image files within using the same process. If the mat file can't be found, it attempts to use a backup .mat file. If that can't be found it uses hardcoded image directories, which also if not found uses the same process as above.