Version: 0.71
Cleaned up a bit of the preference callback code
All of the framebuffer classes the renderer uses needed to resize when the camera resized, duplicating the number of callbacks. So I instead just made 1 of each type of callback in the renderer itself and made it then call all the resize functions etc in all the buffers. Centralized the callbacks in a nutshell. This means the buffers don't need the enginePackage parameter.
Fixed bug with window resizing
Removed old visual fx.