Version: 0.96
Cleaned up a bit of the documentation, fixed a few issues with factories returning gibberish references, and swapped out all maps that used char keys with MappedChar.

MappedChar is an adaptor class that wraps std::map<const char*, T>. Working with maps that use char* keys meant they all neede to implement structs for comparison. Further, it was verbose writing out the insertions, checking for elements... this just makes it a lot neater. Also made a wrapper for it for vector elements of type T (see VectorMap)