Version: 0.108
Overhauled preference callback logic. No longer using a million structs, instead using a better observer pattern taking advantage of c++11. https://juanchopanzacpp.wordpress.com/2013/02/24/simple-observer-pattern-implementation-c11/
The new addPrefCallback works similarly to before, except now it uses a lambda function that gets sent as a parameter to be stored. it also takes in a pointer from the sender to be used as a handle, allowing the sender to easily opt-out of the callback by passing its reference.