Version: 0.117
Started modifying the reflection system to make way for the rest of the reflecting mechanisms that will get implemented.
For now, it temporarily uses the skybox to fill the reflection buffer, and the ssr shader reads from said buffer before applying the effect to the screen.

Also introduced a base class for frame buffers