Version: 0.129
Fixed an issue where models caused crashes when rendering too soon.
Cause: Model asset completed loading, thus returns true on exists. However, it may not have finished (or even begun) calling its callbacks, so VAO's may be incomplete. This can cause a driver crash.
Fixed: Wherever vao's are used, a boolean is updated when the vao is updated within the callback, and won't be rendered until the vao is true.