Version: 0.130
Changed model rendering to use 1 UBO total, where all models share the one with with their own data as elements in the UBO array.

This works for the geometry pass, but for shadows it doesn't yet because the lights themselves don't own the shader.

I will proceed to the next stage where this shouldn't matter, using multiraw