Version: 0.132
I've finally fixed a bug plaguing my model rendering.
The issue was extremely low perf rendering geometry. It appears to be a compiler issue with the geometry shader, so for now I've circumvented it by purely refering to objects in buffers without storing local copies in the code anywhere.

I'm leaving the code cleanup + documentation for next commit