Version: 0.136
Renamed the mapped buffers into the static and dynamic versions, to make it more clear what they do.
Also, implemented a new class "VectorBuffer" that is a template type class, which acts like a dynamic buffer but works on elements like an array.

This patch isn't perfect yet, but I need to save my changes for now.