Version: 0.138
Updated directional lights to use the new buffer technique
The scene is set up to only use 1 directional light, so in theory this technique SHOULD be faster with multiple lights. However, with 1 light we actually have a 2ms loss of performance.
I've updated the shadow shaders to reflect these changes, however I forgot to update the radiance shaders.