Version: 0.153
Not as much changes as the last, but I'm very tired.

Removed the position channel from the shadow FBO and altered shaders to read from the depth buffer instead and then reconstruct the position from the depth map.
This required the depth texture to not have a comparison mode on, meaning we no longer get free PCF from the GPU, and had to modify all shadow-lighting shaders.
Optimized lighting a bit by pre-calculating light P * light V, and sending that to the lighting buffer. Also sending its INVERSE since it is used in indirect lighting.

Indirect point lighting is broken, pending a solution to calculating inverse paraboloid or a replacement matrix.