Version: 1.2
Added the component creator pattern back, and added level loading.

I realized that in order for the level loading to be short and efficient, it needed to be void of component type, delegating the component creation entirely to the ECS.
Its bad enough that the world is responsible for the individual parameter data types, but maybe that can be cleaned up some other day.