Version: 1.9.A
To improve performance, I've figured out how to remove the pbr lighting calculations from the skybox reflector and the local cube reflector, and joined it into the final reflection calculation that was adding the AO value to it.

I've also made some changes to the shaders, debugging to see when discards and branching changed performance.