Version: 1.17
- Shader performance improvements:
Instead of indexing into shader storage buffers in the fragment shaders, I've changed the included shaders to lookup the values in the vertex shaders instead, and pass the value (non-interpolated) to the pixel shader. Even though this involves copying that data, it is faster due to not indexing repeatedly + the data needed is now coherent.

- Removed gamma variable from camera buffer structs, replaced with shader uniform (1 place used only)
- Removed Near, and Far plane variables, Near plane is hardcoded as a define, and Farplane + FOV exist but are the 2 padded out variables (application reads/writes them, shaders ignore them)

Overall, performance has improved A LOT.