Version: 1.24
Models try to import a ".mat" file when loading their material textures.
This isn't a binary operation however, meaning we will still take record of the material paths defined within the model itself, but whatever is defined in the .mat file overwrites whatever was defined within the model.

So if a .mat is present, it takes precedence, falling back on whatever is defined in the model, falling back on default textures.