3.7.3
Changed graphics pipeline such that it has an update pass, which first updates components of all graphics systems, and then calls updateTechnique on them.
Currently, all extra rendering like shadows and reflectors are done in this pre-pass, except for directional shadows. This update pass is done once a frame, not during the pipeline render() call.