3.8.6
Reworked ENTIRE graphics system to better facilitate the ECS approach

List of active viewports updated every frame, shared with techniques that care.
Previously, techniques only cared about the main camera's perspective. Everything that did, now must maintain a vector of that data, where each index corresponds to an index in the shared viewport list. It doesn't matter if the index changes between frames.

Shadows are handled a bit differently.
We don't want to have shadows/reflectors update per viewport, instead prior to deciding which of the oldest are the most applicable to update, we check if it is visible at all from all perspectives instead of just that one perspective. This way, we can update once per frame.